bpf: srv6: extract maps to srv6_maps.h
Some code in srv6.h expects additional configuration macros that the
alignchecker currently doesn't provide. Let's move the maps-related parts
into a standalone header, so that the alignchecker can then only include
this header.
